// src/app/countries/interfaces/cache-storage-interface.ts
import { Country } from "./country";
import { Region } from "./rigeon.type";

export interface CacheStore {
    byCapital: TermCountries;
    byRegion: RegionCountries;
    byCountries: TermCountries;
}

export interface TermCountries {
    term: string;
    countries: Country[];
}

export interface RegionCountries {
    region?: Region;
    countries: Country[];
}